services/transcriber/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
openai-whisper==20231117
librosa==0.10.1
vosk==0.3.45
faster-whisper==0.10.0
numpy

pytest==8.3.3
httpx==0.27.2
